플레이어 탄약 테스트 코드 작성, 조준경 흔들림 버그 수정, 숨참기 테스트 애니메이터 각도, 스케일, 회전 값, 추가 오리진 오프셋 추가 유리병&드럼&원판 생성 애니메이션 추가 총알 발사, 병, 드럼 생성 및 파괴 소리 추가

+++ SFMLMaterialSniper/UiHud.cpp
@@ -106,7 +106,7 @@ void UiHud::SetAmmo(int ammo)
 	textAmmo.SetString("Ammo : " + std::to_string(ammo));
 }
 
-void UiHud::SetBreath(int breath)
+void UiHud::SetBreath(float breath)
 {
 	textBreath.SetString("Breath : " + std::to_string(breath));
 }
